+++ wiki_rag/config_update/__init__.py
@@ -1,0 +1,4 @@
+#  Copyright (c) 2025, Moodle HQ - Research
+#  SPDX-License-Identifier: BSD-3-Clause
+
+"""Config update utility for converting .env to config.yaml."""